Move checkout form handling to request handler extension
