+++ src/Checkout/CheckoutFormExtension.php
@@ -4,14 +4,9 @@ declare(strict_types=1);
 namespace SwipeStripe\Accounts\Checkout;
 
 use SilverStripe\Core\Extension;
-use SilverStripe\Core\Injector\Injector;
 use SilverStripe\Forms\FieldList;
 use SilverStripe\Forms\OptionsetField;
-use SilverStripe\Security\IdentityStore;
-use SilverStripe\Security\Member;
 use SilverStripe\Security\Security;
-use SwipeStripe\Accounts\AccountCreationEmail;
-use SwipeStripe\Accounts\Customer\MemberExtension;
 use SwipeStripe\Accounts\Order\OrderExtension;
 use SwipeStripe\Order\Checkout\CheckoutForm;
 use SwipeStripe\Order\Order;
@@ -28,19 +23,6 @@ class CheckoutFormExtension extends Extension
 
     const CHECKOUT_GUEST = 'Guest';
     const CHECKOUT_CREATE_ACCOUNT = 'Account';
-
-    /**
-     * @var IdentityStore
-     */
-    protected $identityStore;
-
-    /**
-     * CheckoutFormExtension constructor.
-     */
-    public function __construct()
-    {
-        $this->identityStore = Injector::inst()->get(IdentityStore::class);
-    }
 
     /**
      *
@@ -67,42 +49,5 @@ class CheckoutFormExtension extends Extension
         ], static::CHECKOUT_CREATE_ACCOUNT));
 
         $fields->add(CheckoutPasswordField::create(static::ACCOUNT_PASSWORD_FIELD));
-    }
-
-    /**
-     * @param array $data
-     * @throws \SilverStripe\ORM\ValidationException
-     */
-    public function beforeInitPayment(array $data): void
-    {
-        $cart = $this->owner->getCart();
-
-        if (Security::getCurrentUser() === null && $data[static::GUEST_OR_ACCOUNT_FIELD] === static::CHECKOUT_CREATE_ACCOUNT) {
-            $newAccount = $this->createCustomerAccount($cart, $data);
-            $this->identityStore->logIn($newAccount, false, $this->owner->getController()->getRequest());
-            Security::setCurrentUser($newAccount);
-        }
-
-        $cart->MemberID = Security::getCurrentUser() ? Security::getCurrentUser()->ID : 0;
-    }
-
-    /**
-     * @param Order $cart
-     * @param array $data
-     * @return Member
-     */
-    protected function createCustomerAccount(Order $cart, array $data): Member
-    {
-        /** @var Member|MemberExtension $member */
-        $member = Member::create();
-        $member->Email = $data['CustomerEmail'];
-        $member->setName($data['CustomerName']);
-        $member->changePassword($data[static::ACCOUNT_PASSWORD_FIELD]['_Password'], false);
-        $member->DefaultBillingAddress->setValue($cart->BillingAddress);
-
-        $member->setField(AccountCreationEmail::SEND_EMAIL_FLAG, true);
-        $member->write();
-
-        return $member;
     }
 }
